fix(cppcheck): clear remaining perf/portability/style findings (#9)

+++ src/graph-stepguider.cpp
@@ -97,7 +97,7 @@ GraphStepguiderWindow::GraphStepguiderWindow(wxWindow *parent)
     pMainSizer->Add(pLeftSizer, wxSizerFlags().Expand());
 
     m_pClient->m_length = pConfig->Global.GetInt("/graph_stepguider/length", 1);
-    wxString label = wxString::Format(_T("%3d"), m_pClient->m_length);
+    wxString label = wxString::Format(_T("%3d"), (int) m_pClient->m_length);
     LengthButton =
         new OptionsButton(this, BUTTON_GRAPH_LENGTH, label, wxPoint(10, 10), wxSize(80, -1), wxALIGN_CENTER_HORIZONTAL);
     LengthButton->SetToolTip(_("Select the number of frames of history to display"));
@@ -135,7 +135,7 @@ void GraphStepguiderWindow::OnButtonLength(wxCommandEvent& WXUNUSED(evt))
     unsigned int val = 1;
     for (int id = MENU_LENGTH_BEGIN; id <= MENU_LENGTH_END; id++)
     {
-        wxMenuItem *item = menu->AppendRadioItem(id, wxString::Format("%d", val));
+        wxMenuItem *item = menu->AppendRadioItem(id, wxString::Format("%d", (int) val));
         if (val == m_pClient->m_length)
             item->Check(true);
         val *= 4;
@@ -156,7 +156,7 @@ void GraphStepguiderWindow::OnMenuLength(wxCommandEvent& evt)
 
     m_pClient->m_length = val;
 
-    LengthButton->SetLabel(wxString::Format(_T("%3d"), val));
+    LengthButton->SetLabel(wxString::Format(_T("%3d"), (int) val));
 
     pConfig->Global.SetInt("/graph_stepguider/length", val);
